+++ packages/dumb/src/modules/plot-methods/charts/bolus-carbs/meal.ts
@@ -29,16 +29,16 @@ import _ from 'lodash'
 import * as d3 from 'd3'
 
 import { type Wizard } from 'medical-domain'
-import { Pool } from '../../../models/pool.model'
+import { Pool } from '../../../../models/pool.model'
 import {
   drawCircle,
   drawText,
   drawVerticalRectangle,
   getTooltipContainer
-} from '../../../utils/daily-chart/daily-chart.util'
-import { PlotFunction } from '../../../models/plot-function.model'
-import { PlotSelection } from '../../../models/plot-selection.model'
-import { PlotOptions } from '../../../models/plot-options.model'
+} from '../../../../utils/daily-chart/daily-chart.util'
+import { PlotFunction } from '../../../../models/plot-function.model'
+import { PlotSelection } from '../../../../models/plot-selection.model'
+import { PlotOptions } from '../../../../models/plot-options.model'
 import {
   getBolusFromInsulinEvent,
   getBolusType,
@@ -49,8 +49,8 @@ import {
   isMealWithOverride,
   isMealWithUndelivered,
   isMealWithUnderride
-} from '../../../utils/bolus/bolus.util'
-import { BolusType } from '../../../models/enums/bolus-type.enum'
+} from '../../../../utils/bolus/bolus.util'
+import { BolusType } from '../../../../models/enums/bolus-type.enum'
 
 const BOLUS_WIDTH = 12
 const HALF_BOLUS_WIDTH = BOLUS_WIDTH / 2