+++ packages/dumb/src/modules/plot-methods/charts/common/confidential-mode.ts
@@ -31,7 +31,7 @@ import * as d3 from 'd3'
 import lockIcon from 'lock.svg'
 
 import { type ConfidentialMode } from 'medical-domain'
-import { type Pool } from '../../../models/pool.model'
+import { type Pool } from '../../../../models/pool.model'
 import {
   calculateWidth,
   drawImage,
@@ -39,12 +39,12 @@ import {
   drawZoneRectangle,
   getTooltipContainer,
   xPos
-} from '../../../utils/daily-chart/daily-chart.util'
-import { type PlotFunction } from '../../../models/plot-function.model'
-import { type PlotSelection } from '../../../models/plot-selection.model'
-import { type PlotOptions } from '../../../models/plot-options.model'
-import { createIdGenerator } from '../../../utils/id-generator/id-generator.util'
-import { DailyPlotElement } from '../../../models/enums/daily-plot-element.enum'
+} from '../../../../utils/daily-chart/daily-chart.util'
+import { type PlotFunction } from '../../../../models/plot-function.model'
+import { type PlotSelection } from '../../../../models/plot-selection.model'
+import { type PlotOptions } from '../../../../models/plot-options.model'
+import { createIdGenerator } from '../../../../utils/id-generator/id-generator.util'
+import { DailyPlotElement } from '../../../../models/enums/daily-plot-element.enum'
 
 // ID generator for consistent element identification
 const idGen = createIdGenerator(DailyPlotElement.ConfidentialMode)