+++ packages/dumb/src/modules/plot-methods/charts/events/night-mode.ts
@@ -30,19 +30,19 @@ import * as d3 from 'd3'
 import nightModeIcon from 'night-mode.svg'
 
 import { type NightMode } from 'medical-domain'
-import { type Pool } from '../../../models/pool.model'
+import { type Pool } from '../../../../models/pool.model'
 import {
   calculateWidth,
   drawImage,
   drawZoneRectangle,
   getTooltipContainer,
   xPos
-} from '../../../utils/daily-chart/daily-chart.util'
-import { type PlotFunction } from '../../../models/plot-function.model'
-import { type PlotSelection } from '../../../models/plot-selection.model'
-import { type PlotOptions } from '../../../models/plot-options.model'
-import { createIdGenerator } from '../../../utils/id-generator/id-generator.util'
-import { DailyPlotElement } from '../../../models/enums/daily-plot-element.enum'
+} from '../../../../utils/daily-chart/daily-chart.util'
+import { type PlotFunction } from '../../../../models/plot-function.model'
+import { type PlotSelection } from '../../../../models/plot-selection.model'
+import { type PlotOptions } from '../../../../models/plot-options.model'
+import { createIdGenerator } from '../../../../utils/id-generator/id-generator.util'
+import { DailyPlotElement } from '../../../../models/enums/daily-plot-element.enum'
 
 // ID generator for consistent element identification
 const idGen = createIdGenerator(DailyPlotElement.NightMode)