+++ packages/dumb/src/modules/plot-methods/charts/events/physical-activity.ts
@@ -30,21 +30,21 @@ import * as d3 from 'd3'
 import physicalActivityIcon from 'physical-activity.png'
 
 import { type PhysicalActivity } from 'medical-domain'
-import { type Pool } from '../../../models/pool.model'
+import { type Pool } from '../../../../models/pool.model'
 import {
   calculateWidth,
   drawImage,
   drawZoneRectangle,
   getTooltipContainer,
   xPos
-} from '../../../utils/daily-chart/daily-chart.util'
-import { type PlotFunction } from '../../../models/plot-function.model'
-import { type PlotSelection } from '../../../models/plot-selection.model'
-import { type PlotOptions } from '../../../models/plot-options.model'
-import { createIdGenerator } from '../../../utils/id-generator/id-generator.util'
+} from '../../../../utils/daily-chart/daily-chart.util'
+import { type PlotFunction } from '../../../../models/plot-function.model'
+import { type PlotSelection } from '../../../../models/plot-selection.model'
+import { type PlotOptions } from '../../../../models/plot-options.model'
+import { createIdGenerator } from '../../../../utils/id-generator/id-generator.util'
 
-import { DailyPlotElement } from '../../../models/enums/daily-plot-element.enum'
-import { PLOT_DIMENSIONS } from '../../../models/constants/plot.constants'
+import { DailyPlotElement } from '../../../../models/enums/daily-plot-element.enum'
+import { PLOT_DIMENSIONS } from '../../../../models/constants/plot.constants'
 
 // ID generator for consistent element identification
 const idGen = createIdGenerator(DailyPlotElement.PhysicalActivity)